Misc changes, primarily switching away from int_writer, as I can no longer reproduce the 5 percent speed up after switching to tmpfs.
We now split of the packets into one file, instead of 10k smaller ones.
tmpfs's blocksize is 4k, so files less than 4k bytes large end up being stored
in 4k bytes. This was causing 4mb's worth of data to take over 50mb of
ram, which is painful to handle, to say the least.

Since we store them all one one file, to make handling the separate
packets easier, we align the section starts to 16 bytes.

Signed-off-by: Hal Gentz <zegentzy@protonmail.com>